moved forgot password logic to controller
