+++ controllers/user.controller.js
@@ -1,7 +1,8 @@
 const httpStatus = require('http-status');
 const User = require('../models/user.model');
 const sendResponse = require('../helpers/response');
-const nodemailer = require('nodemailer');
+const PasswordReset = require('../models/passwordReset.model');
+const sendMail = require('../helpers/SendMail');
 
 /**
  * Load user and append to req.
@@ -21,25 +22,47 @@ exports.getUsers = (req, res) => {
   return res.json(sendResponse(200, 'testing', null, null));
 };
 
-exports.sendMail = (email, message) => {
-  const smtpTransport = nodemailer.createTransport({
-    service: 'gmail',
-    secure: true,
-    port: 25,
-    auth: {
-      user: process.env.MAILER_EMAIL,
-      pass: process.env.MAILER_PASSWORD
-    }
+exports.forgotPassword = async (req, res) => {
+  const user = await User.getByEmail(req.body.email);
+  if (!user) {
+    return res.json(
+      sendResponse(httpStatus.NOT_FOUND, 'User not found', null, null, null)
+    );
+  }
+  const passwordReset = new PasswordReset({
+    userID: user._id,
+    email: req.body.email,
+    isAdmin: user.isAdmin
   });
-  let mailOptions = {
-    to: email,
-    from: 'passwordreset@mentordev.com',
-    subject: 'Mentor Dev Password Reset',
-    text: message
-  };
-  smtpTransport.sendMail(mailOptions, err => {
-    if (err) {
-      return new Error(err);
-    }
-  });
+  const passwordResetResult = await passwordReset.save();
+  if (!passwordResetResult || !passwordResetResult.email) {
+    return res.json(
+      sendResponse(
+        httpStatus.NOT_FOUND,
+        'something went wrong',
+        null,
+        null,
+        null
+      )
+    );
+  }
+  const message = `You are receiving this because you (or someone else) have requested the reset of the password for your account.
+  
+  Please click on the following link, or paste this into your browser to complete the process:
+  
+  http://${req.headers.host}/reset/${passwordResetResult.resetPasswordToken}
+  
+  If you did not request this, please ignore this email and your password will remain unchanged`;
+
+  sendMail(passwordResetResult.email, message);
+
+  return res.json(
+    sendResponse(
+      httpStatus.OK,
+      'Successfully sent reset mail',
+      null,
+      null,
+      null
+    )
+  );
 };
